Remove pseudo array object type from the project
A new class type enum is introduced to describe the class of objects.
This enum is organized to improve property resolve and GC performance.

JerryScript-DCO-1.0-Signed-off-by: Zoltan Herczeg zherczeg.u-szeged@partner.samsung.com

+++ jerry-core/ecma/operations/ecma-promise-object.h
@@ -98,7 +98,7 @@ typedef struct
 bool ecma_is_promise (ecma_object_t *obj_p);
 ecma_value_t ecma_op_create_promise_object (ecma_value_t executor, ecma_value_t parent,
                                             ecma_object_t *new_target_p);
-uint16_t ecma_promise_get_flags (ecma_object_t *promise_p);
+uint8_t ecma_promise_get_flags (ecma_object_t *promise_p);
 ecma_value_t ecma_promise_get_result (ecma_object_t *promise_p);
 void ecma_reject_promise (ecma_value_t promise, ecma_value_t reason);
 void ecma_fulfill_promise (ecma_value_t promise, ecma_value_t value);
